refactor: remove sub folder in marketing components
