+++ apps/trpkit.com/src/app/(marketing)/page.tsx
@@ -1,5 +1,5 @@
-import Hero from "@/components/marketing/home/Hero";
-import WebsiteMonitoring from "@/components/marketing/home/WebsiteMonitoring";
+import Hero from "@/components/marketing/Hero";
+import WebsiteMonitoring from "@/components/marketing/WebsiteMonitoring";
 import type React from "react";
 
 export default function Page() {
